Implement Heat Pump Only Mode
If the resistance elements are on, turn off, and engage compressor. Logic is set to say enegage compressor BUT if the compressor gets locked out, i.e. below the min operating temperature  for the compressor, use the backup resistance elements, we think this is how it works in reality. But after this example of heating happens the compressor will be engaged, the backup resistance is disengaged, but the compressor is locked out and did not heat, and the backup resistance has added heat to the system.

Made new test